Fix sidebar terminal deadlock and persist-before-shutdown race
Fix two bugs found via e2e test failures:

1. Reentrant mutex deadlock in sidebar terminal response writer:
   SidebarPTYFlush handler holds ts.mu, calls VTerm.Write(), which
   triggers the response writer callback that tried to acquire ts.mu
   again. Fixed by capturing the terminal reference directly instead
   of locking through ts.mu.

2. Tabs persisted as "stopped" on quit: Shutdown() closes tabs (sets
   Running=false) before the 500ms persist debounce fires, so tabs
   were saved with "stopped" status and skipped on restore. Fixed by
   adding persistAllWorkspacesNow() called synchronously before
   Shutdown() in the quit handler.

Extract persistence functions to app_persistence.go to keep
app_input_messages.go under the 500-line limit.
