+++ internal/app/app_input_dialogs.go
@@ -99,6 +99,10 @@ func (a *App) handleDialogResult(result common.DialogResult) tea.Cmd {
 		}
 
 	case DialogQuit:
+		// Persist workspace tabs synchronously before shutdown.
+		// Shutdown() closes tabs (sets Running=false), so we must
+		// capture current state first to avoid saving "stopped" status.
+		a.persistAllWorkspacesNow()
 		a.Shutdown()
 		a.quitting = true
 		return tea.Quit
